+++ HISTORY.rst
@@ -20,6 +20,9 @@ Latest (Unreleased)
 +++++++++++++++++++
 
 **Features and Improvements**
+
+* BSIBSO-1125: Modification of the purchase_id domain in the supplier invoice
+  to have only purchase orders in the currency of the invoice
 * BSIBSO-1121: Active field in 'purchase.order', filter into view
 * BSIBSO-1117: have a customer dependent numbering for analytic accounts